+++ blueprints/flexberry-model/files/__root__/mixins/regenerated/models/__name__.js
@@ -4,13 +4,14 @@ import DS from 'ember-data';
 export let Model = Ember.Mixin.create({
 <%= model %>
 });
-
+<% if (namespace) { %>
 export function defineNamespace(modelClass) {
   modelClass.reopenClass({
     namespace: '<%= namespace %>',
   });
 }
-<%if(parentModelName) {%>
+<% }
+if(parentModelName) {%>
 export let defineBaseModel = function (modelClass) {
   modelClass.reopenClass({
     _parentModelName: '<%= parentModelName %>'