+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-* The `flexberry-model` blueprint generates the exported function to define the namespace of the model.
-* The `flexberry-model-init` blueprint generates an import and call of function to define the namespace in the model file.
+* The `flexberry-model` blueprint generates the exported function to define the namespace of the model, if it is not empty.
+* The `flexberry-model-init` blueprint generates an import and call of function to define the namespace in the model file, if it is not empty.
 
 ## [0.12.1] - 2018-08-22
 ### Added